pyproject.toml
[project]
name = "knozify-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-phonenumber-field[phonenumbers]>=8.1.0",
    "django-redis>=5.4.0",
    "django-split-settings>=1.3.2",
    "django-watson>=1.6.3",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "google-auth>=2.40.1",
    "logtail-python>=0.3.3",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.5",
    "pytest>=8.4.0",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.1.0",
    "sentry-sdk[django]>=2.27.0",
    "social-auth-app-django>=5.4.3",
]

[dependency-groups]
dev = [
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "pre-commit>=4.2.0",
]
